Create acarajé recipe page
Create a new directory within the odin-recipes directory and name it
recipes.

Create acaraje.html within recipes directory, which is the name of the
first recipe.

Include a h1 heading with the recipe's name.

Back in the index.html and add a link to the recipe page that was just
created.